requirements.txt
anytree==2.13.0
camelot-py==1.0.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
commonmark==0.9.1
cryptography==45.0.4
et_xmlfile==2.0.0
ghostscript==0.8.1
numpy==2.2.6
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
pandas==2.3.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pycparser==2.22
Pygments==2.19.1
PyMuPDF==1.26.0
pypdf==5.6.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-dotenv==0.21.1
pytz==2025.2
rich==12.6.0
rich_argparse_plus==0.3.1.4
setuptools==78.1.1
six==1.17.0
tabulate==0.9.0
tzdata==2025.2
vulture==2.14
wheel==0.45.1
yara-python==4.5.4
yaralyzer==0.9.6
